requirements.txt
flask
pymupdf
python-docx
google-generativeai